feat: replace rppal dependency with embedded-hal 1.x
Replaces the hard dependency on rppal with the generic embedded-hal traits. This
allows to use this driver with a wide array of hardware, as long as the HAL
implementation supports embedded-hal 1.x.

Closes #1

BREAKING CHANGE: You now need to supply your own i2c adapter, either from rppal
or from linux-embedded-hal.